+++ src/app/api/image-urls/route.ts
@@ -1,7 +1,6 @@
-// import type { NextApiRequest, NextApiResponse } from "next";
 import { createClient } from '@supabase/supabase-js'
 import { NextRequest, NextResponse } from "next/server";
-// import type { NextRequest, NextResponse } from 'next/server';
+import { getImageUrls } from '@/lib/supabaseDb/supabase';
 
 console.log('Entering api/image-urls/route.ts')
 
@@ -10,36 +9,8 @@ console.log('Entering api/image-urls/route.ts')
 export async function GET(req: NextRequest) {
 
   console.log('getting Image URLs...')
-  // console.log('getHomepageImageUrls/request: ', req)
-  // const cookieStore = cookies();
-  const { SUPABASE_URL, SUPABASE_KEY } =
-    process.env;
-
-  const supabase = createClient(SUPABASE_URL!, SUPABASE_KEY!, {
-  })
-
   try {
-    // get images from supabase;
-    const { data, error } = await supabase
-      .storage
-      .from('products')
-      .list('arrangements', {
-        limit: 10,
-        offset: 1
-      })
-
-    // console.log('api/image-urls/route.ts/GET/data: ', data)
-
-    const urls: string[] | undefined = data?.map((img) => {
-      const url = supabase.storage
-        .from('products/arrangements')
-        .getPublicUrl(img.name);
-      // console.log('serverActions/getImages/urls/map/url: ', url)
-      return url.data.publicUrl;
-      // return img.name
-    })
-    console.log('/api/image-urls/route.ts/GET/urls: ', urls)
-
+    const urls = await getImageUrls();
     return NextResponse.json(urls);
 
   }
