+++ src/lib/mongoDb/mongo.ts
@@ -1,0 +1,14 @@
+import mongoose from "mongoose";
+import { Admin, User } from "./models";
+
+export async function createAdmin() {
+
+
+
+}
+
+export async function createUser() {
+
+
+
+}
